# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: klimatsearch
  version: 0.1.0
  description: |
    Hybrid search HTTP API for [Boverket Klimatdatabas](https://www.boverket.se/sv/klimatdeklaration/klimatdatabas/).
    Same binary also serves MCP at `/mcp` (streamable) and `/mcp/sse` (legacy).

    Cite **Boverket Klimatdatabas**. This project is not affiliated with Boverket.

    Boverket's OpenAPI has no per-id GET. Each Boverket Resource includes `origin`
    (`https://klimatdatabasen.boverket.se/detaljer/{category_code}/{id}`).
    `GET /api/resources/{id}/origin` 302s there.

    Resources include typical A1–A3 (`a1a3`) plus conservative A1–A3, A4, A5.1,
    waste/conservative factors, biogenic carbon, service life, BK04, and A4
    transport legs when Boverket publishes them. Energy carriers omit A4/A5.1.

    `GET /healthz`, `GET /openapi.yaml`, and `GET /docs` are public.
    The default self-hosted build (`make build`) has no Unkey or MPP and does not
    gate `/api/*`. The binary we run is `make build-hosted` (`-tags hosted`).

    Outbound: when ingest upserts a changed Resource, klimatsearch POSTs `catalog.changed`
    to `KLIMAT_WEBHOOK_URL`. That is not an inbound path on this API.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: klimatsearch
    url: https://github.com/mong-x/klimatsearch
...
servers:
  - url: http://127.0.0.1:8080
    description: make run default
  - url: http://127.0.0.1:8081
...
paths:
...
  /api/resources/{id}:
    get:
      tags: [Resources]
      summary: Get one Resource
      description: |
        `{id}` may be prefixed (`boverket:6000000000`) or a bare Resource ID
        when only one Catalog matches (409 if ambiguous).
      operationId: getResource
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Full Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    unkey:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Unkey keyspace API key (`sk_…`). Not the server root key.
    mpp:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Machine Payments Protocol. Value is `Payment <credential>`.
        Missing credential → 402 with `WWW-Authenticate: Payment`.
  parameters:
...
    Lang:
      name: lang
      in: query
      required: false
      schema:
        type: string
        enum: [sv, en]
        default: sv
...
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Prefixed `catalog:id` or bare Resource ID
      example: boverket:6000000000
  responses:
    Error:
      description: Error object
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Guard denied (no MPP / invalid Bearer)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PaymentRequired:
      description: MPP challenge
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Payment realm="klimatsearch"
      content:
        application/problem+json:
          schema:
            type: object
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Resource:
      type: object
      required: [id, catalog, source]
      properties:
        id:
          type: string
          example: "6000000000"
        catalog:
          type: string
          enum: [boverket, br25]
        name_sv:
          type: string
        name_en:
          type: string
        description_sv:
          type: string
        description_en:
          type: string
        applicability_sv:
          type: string
        applicability_en:
          type: string
        synonyms:
          type: string
        a1a3:
          type: number
          format: double
          description: Typical GWP-GHG A1–A3 kg CO2e per Declared unit (comparison / typical value, not the conservative declaration factor)
        a1a3_conservative:
          type: number
          format: double
          description: Conservative A1–A3 (~25% above typical). Boverket uses this for climate declarations; klimatsearch Compare uses typical.
        a4:
          type: number
          format: double
          description: GWP-GHG module A4 (transport to site). Omitted for energy carriers.
        a5_1:
          type: number
          format: double
          description: GWP-GHG module A5.1 (construction-site waste). Omitted for energy carriers.
        gwp_unit:
          type: string
          example: kg CO2 eq./kg
        conservative_factor:
          type: number
          format: double
        waste_factor:
          type: number
          format: double
        biogenic_carbon:
          type: number
          format: double
        service_life:
          type: string
        transports:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              distance_km:
                type: number
              type:
                type: string
              fuel:
                type: string
              fuel_resource_id:
                type: string
        unit:
          type: string
          example: kg
        conversions:
          type: object
          additionalProperties:
            type: number
            format: double
        category:
          type: string
        version:
          type: string
          example: "02.07.000"
        source:
          type: string
          example: Boverket Klimatdatabas
        origin:
          type: string
          format: uri
          description: Official Boverket product sheet when catalog is boverket
          example: https://klimatdatabasen.boverket.se/detaljer/10/6000000000
...
security:
  - {}
  - unkey: []
  - mpp: []